GEODE-7921: NullPointerExceptions logged during auto-reconnect (#4898)
Don't deliver cache-level messages that were queued while disconnected
for auto-reconnect.  During auto-reconnect there is a QuorumChecker that
receives messages on the jgroups channel and tries to establish
communications with a quorum of the old membership view.  It may also
get JoinRequest messages and other membership-level messages but I
observed one case where it also queued cache-level messages when the
property disable-tcp was set to true (funnelling all comms through
jgroups).

I also added some null checks to the LatestLastAccessTimeMessage and a
small test for that.