GEODE-7852: test ClientHealthMonitor functionality behind a SNI gateway
This ensures that a server sitting behind an SNI gateway detects the
loss of a client and cleans up after it.  In this case the test detects
that the server has closed CQs created by the non-durable client.

Since test code is not accessible in the Docker container that's running
the server I've enhanced the StatArchiveReader to be able to report the
values of a statistic and have enabled statistics recording in the
server.

(cherry picked from commit 376df4cc7abc6db38b4b298f8af6f3c53baedf15)