GEODE-8020: buffer corruption in SSL communications (#4994)
revert change in GEODE-6661 that made NioSslEngine use a direct buffer.
This class is not designed to share its buffer with a pool in
BufferPool.  Connection is also modified to use a heap buffer for
reading encrypted SSL packets for consistency.  New tests ensure that
these buffers are the correct type when using SSL or plain sockets.