GEODE-7565: Allow gateway receivers with same host and port (2nd try)
This reverts the revert of the original commit, adds extra tests and
also fixes the originally introduced issues.

----

There was a problem with Geode WAN replication when GW receivers are
configured with the same hostname-for-senders and port on all servers.
The reason for such a setup is deploying Geode cluster on a Kubernetes
cluster where all GW receivers are reachable from the outside world on
the same VIP and port.

The problem experienced is that shutting down one server is stopping
replication to this cluster until the server is up again. This is
because Geode incorrectly assumes there are no more alive servers when
just one of them is down, because since they share hostname-for-senders
and port, they are treated as one same server.

With these changes locator is able to distinguish the different
receivers using the same hostname and port so replication is not
impacted when one server is stopped.