GEODE-8070: rework handling of "any" in SSLUtils
Rewrite of the sslContext location method to fix these problems:
  requesting a specific but non-existent protocol would not result in an
  exception

  requesting "any" (the default) would result in a less-secure protocol
  than is actually available (e.g., SSL instead of TLSv1.2)

I also changed processing so that if "any" is in the list but none of
the default protocols are available we continue to search through the
provided list.  For instance, {X, any, Y} would result in a search for Y
if X and all of the default protocols cannot be found.