GEODE-8029: Delete orphaned drf files (2nd try)
This reverts the revert of the original commit and also fixes the
originally introduced issues.

----

The OpLog initialization now delete unused drf files to prevent the
proliferation of unused records and files within the system, which
can could cause members to fail during startup while recovering
disk-stores (especially when they are isolated for gateway-senders).

- Added distributed tests.
- Delete orphaned drf files when deleting the corresponding
  crf during recovery.