GEODE-8131: reader thread blocked attempting to issue an alert
This PR does not solve the problem of the alert system causing a P2P
message reader to block but instead ensures that the reader thread will
write deserialization information to the log before issuing a fatal
alert.  If the alert level is set to "info" this won't help, but no-one
would set the alert level to that level.

Along the way I removed some duplicate strings.