GEODE-8241: Locator observes locator-wait-time (#5236)
In the case where a locator starts up and is unable to connect to any
other locators, it may decide to become the membership coordinator even
if locator-wait-time has not elapsed.

This change addresses this issue by requiring a locator to wait for
locator-wait-time before deciding to become the coordinator.

Co-authored-by: Aaron Lindsey <alindsey@vmware.com>
Co-authored-by: Vincent Ford <vford@pivotal.io>
Co-authored-by: Bill Burcham <bburcham@pivotal.io>
(cherry picked from commit 720a4caea2ddb22296aa3225fc5264d2096cdf20)