Reintroduce use of SSLSocket in cluster communications
Use of SSLEngine with NIO has proven to be less robust than using
"old-io" SSLSockets.  This PR reintroduces SSLSockets into cluster
communications and removes the use of "new IO" SSLEngine as an option.

The structure of the org.apache.geode.internal.tcp package is left
intact.  None of the old code including Connection.runOioReader() has
been brought back.  Code differences between use of NIO and OIO are
minimal and are mostly at the socket-read/socket-write level.  The
MsgReader class is modified to handle buffer manipulation that was in
the previous "old IO" implementation but was delegated to the "filter"
implementation prior to this PR.

Performance testing has shown that serialization and transmission of
messages works best with "direct" memory buffers while the reading of
messages with SSL functions best with "heap" memory buffers.
DirectChannel and DirectReplyProcessor reflect the choice of "direct"
memory buffers for writing while ClusterConnection's inputBuffer reflects
the choice of "heap" memory buffers for reading.

There are a few changes to the NIO "engine" implementations that I made
while testing performance that I decided to keep because they allow the
use of NioPlainEngine with SSLSockets.  That could be an advantage when
performing similar work in client/server communications.

The cluster communications class has been renamed to ClusterConnection
because I'm very tired of the client/server Connection class having the
same name.  Every time I try to open the class with IntelliJ it
recommends the client/server class as the first option.  No more!