GEODE-8391: Skip assertion if unsupported by host
An assertion in AvailablePortJUnitTest.testIsPortAvailable() assumes
that the InetAddress for the local host is not a loopback address.
Though this assumption holds in our CI containers, it fails on some
developer's Macs, and perhaps on other hosts.

This commit changes the test to skip that assertion if the local host
address is a loopback address.

Authored-by: Dale Emery <demery@vmware.com>