GEODE-8407: MergeLogFiles fails to include files with the same name b… (#5428)
* GEODE-8407: MergeLogFiles fails to include files with the same name but in different directories

I modified the upper levels of the merge process to hold a new object
containing the file's display name and its inputStream.  The
logFiles Map now holds the full path of a file and one of these new
objects.  The new test then uses a dirCount of zero to test that
two files with the same name but different parent directories are
properly stored in the Map for processing.

* fix stresstest failure