DUnit session gets single level dir
MemberStarterRule wants to clean up after any members it creates. To do
that, at the end of each test it deletes almost every item remaining in its
current working directory. The only item it refrains from deleting is a
directory named "dunit", which it knows it does not own.

My previous commit violated MemberStarterRule's assumptions by tucking
the "dunit" directory for each session under a temporary directory with
a random name. Because MemberStarterRule did not recongize the random
names, it deleted the temporary directories owned by DUnitLauncher.

This commit prefixes DUnitLauncher's temporary directories with
"dunit-session-" and teaches MemberStarterRule to refrain from deleting
directories so named.

This is a stop-gap solution. Ultimately, the MemberStarterRule in the
TestWorker JVM must learn to create members inside a temporary
directory. This can be either the directory created by DUnitLauncher, or
a temporary directory that the MemberStarterRule creates itself.

Authored-by: Dale Emery <demery@vmware.com>