GEODE-8491: Do not store dropped events in stopped primary gateway sender when possible
Instead of storing dropped events in tmpDroppedEvents to later send batch
removal messages when the primary gateway sender is not started, try to send
the batch removal message when the event to be dropped is received.
That way, when the sender is stopped for a long time and there are events
coming, the memory of the AbstractGatewaySender will not grow with entries
in the tmpDroppedEvents member.

In order to send the batch removal message directly, the eventProcessor for
the AbstractGatewaySender must have been created.
If it is not yet created because the sender was created with manual start
set to true, while receiving events to be dropped, they will be stored in
tmpDroppedEvents as there is no other choice. Nevertheless,
in order to consume less memory, the event stored could be a simplified event
containing only the necessary information to handle it.