GEODE-8683: Honor maximum-time-between-pings in gateway receiver
The maximum-time-between-pings set when creating a gateway receiver
was not honored because the ClientHealthMonitor, that is the singleton
class monitoring all clients supported just one value for maximum
time between pings for all clients. This value is set
when the server in which the receiver is running is started
and when the gateway receiver provides a different value
it is ignored.

With this fix, it is allowed to have different values
for maximum-time-between-clients for different clients
as the value is not taken from the ClientHealthMonitor.
Each client can have its own value.