GEODE-8656: Fix ping only sent to one gateway receiver when several r… (#5670)
* GEODE-8656: Fix ping only sent to one gateway receiver when several receivers share the same hostname-for-senders

When serveral receivers have the same hostname-for-senders only one of them received
pings.
The reason was that the structure holding the endpoints was a Map whose key was
the server location. As all the endpoints for the remote gateway senders shared
the same server location, only one received the ping.

The structure holding endpoints in the endpoint manager has been updated to use
as key a ServerLocationAndMemberId class instead of ServerLocation.