GEODE-7489: DistributionArchUnitTest is running out of memory for some users (#5783)
CoreOnlyUsesMembershipAPIArchUnitTest already restricts use of the
membership API to certain classes and does so in a way that avoids using
too much memory.  Consequently we're removing this test that needs
upwards of 1.5gb of heap since it performs nearly the same test but
for a specific class in the API package instead of the whole package.