GEODE-8765: Fix NullPointerException when group-transaction-events and events in and not in transactions are sent.
A NullPointerException (caught later in the code) was
generated when events without a transactionId in
gateway sender queues were retrieved to complete
transactions.
The problem was that equals was invoked on a
possibly null object (the transactionId).
The call has been rewritten to avoid the
NullPointer exception.

The NullPointerException had different undesired effects
in the Parallel and in the Serial Gateway senders.

In Parallel Gateway Senders it can provoke that events
are left in the queue without ever being drained
although all events are replicated.
In Serial Gateway Senders, the effect is much
more severe as once the NullPointerException is
reached, no events from the queue are replicated.