GEODE-8667: Duplicate online Oplog compaction after offline Oplog compaction (#5689)
After offline compaction, the new Oplog totalCount is zero,
and totalLiveCount is greater than zero, there is no need for further compaction during restart.

When Oplog totalCount is zero, and totalLiveCount is less than or equals to zero,
which indicates an empty Oplog, then compaction is needed.

Update totalCount during recovery, if totalCount == 0 and totalLiveCount > 0, which indicates offline
compaction is done for this Oplog. Update totalCount to the value of totalLiveCount to avoid duplicate compaction.