GEODE-8767: NullPointerException in TCPConduit.getBufferPool due to conTable being null
A NPE was being thrown in TCPConduit.getBufferPool() because the conTable instance
variable is nulled out when the TCPConduit is stopped.  This commit
moves the buffer pool from TCPConduit's conTable object up to the enclosing
DirectChannel object.  This removes the need for DirectChannel to ask
TCPConduit for the buffer pool.