Fixing a suspect string and removing an missed packages element
Fixing a suspect string from RestartOfMemberDistributedTest. This test is
crashing the distributed system, so it is expected to see AlertAppender errors.
We suspect the log4j changes affected  the timing of the test. The fact that we
are seeing AlertAppender errors indicates that log4j is definitely finding our
log4j plugins!

Removing a packages element from the log4j2.xml file checked into the
membership module.