GEODE-8922: Remove ProductUseLog
This removes the ProductUseLog class and its tests.  This class was used
in the Locator to write to a locatorviews.log file showing membership
changes and a little about server load.  All of this information is
already available in statistics and regular log files.  We have never
used this locatorviews.log file as an aid in debugging problems.