GEODE-8971: Add grace period when stopping gateway sender with group-transaction-events enabled
When the gateway sender is stopped there is a high probability that batches with
incomplete transactions are sent even if group-transaction-events is enabled.

The reason is that once the stop command reaches the gateway sender, it immediately
stops queueing events, and this could happen in the middle of receiving events for the
same transaction. If this is the case, some events for the transaction may have
reached the queue right before the stop command was received and the rest of events
for that transaction would not make it to the queue (they would be dropped) because
they arrived right after the stop command was received at the gateway sender.

One way to solve this issue to allow for a grace period when the gateway sender
is stopped during which only events that belong to transactions already
in the queue are queued. Any other event received by the gateway sender
during this grace period would be dropped. After this grace period
the gateway sender will be effectively stopped.