GEODE-8963: separate client/server compatibility from server/server version compatibility
A client's version is used for deserializing data received from the client and for
serializing data sent to the client. It is also used to locate the map of Commands
used to process client requests. Every time we cut a new release we bump this version
in KnownVersions and create a new map of Commands, even though client/server
communications protocols rarely change.

This PR changes KnownVersions to hold a client/server protocol version that is
used to identify clients for command-table selection and serialization rather than
the client's CURRENT_VERSION.