GEODE-8979: CI Failure: SSLSocketHostNameVerificationIntegrationTest (#6079)
This test was closing a client socket before ensuring that a thread it
had created was finished.  If the socket is closed quickly enough it
could cause that thread to get an IOException and cause the test to
fail.

The fix is to ensure that the thread is finished before closing the
client socket.