GEODE-9223: Removing ByteArrayWrapper from hashes
No longer keeping ByteArrayWrapper classes in memory for RedisHashes or
creating ByteArrayWrapper classes when performing hash operations.

The main change to the RedisHash data structure was to switch from a
HashMap<ByteArrayWrapper, ByteArrayWrapper> to a
Object2ObjectOpenCustomHashMap<byte[], byte[]>. This saves ~50 bytes for each
field in the hash to not have the wrappers. Removing the extra allocations when
performing hash operations also helps performance.

Also removing a couple of extra unneeded allocations during logging and message
parsing. We should always create ArrayLists, etc. with a size if it is known so
those objects don't have to resize during parsing/deserialization.

Moving the hash commands executor to a shared object rather than allocating it
for each operation, which is extra GC pressure we don't need.

These changes BREAK compatibility with geode 1.14. The serialization format for
RedisHash, AddsDeltaInfo, and RemsDeltaInfo has changed to not write
ByteArrayWrapper classes on the wire.