GEODE-9080: Add new stackdriver agent to Windows compute images
As a precursor to right-sizing our Windows instances, we need to know
their resource use. The jobs are not long-lived enough for the GCP
Recommendation system to tell us, so we need our own monitoring. Thus,
add stackdriver, so that we can see the data for memory and details
about CPU use.