GEODE-9069: increase timeouts for HExistsDUnitTest (#6212)
the timeout for these DUnit tests were set to 1 second. because these
DUnit tests are run through CI, they can have noisy neighbors that can
cause the condition to time out. This test should not be sensitive to
timing like that because DUnits are not the place for performance
testing.