GEODE-9052: rename the geode-redis module (#6211)
* GEODE-9052: rename the geode-redis module

rename the geode-redis module to geode-apis-compatible-with-redis. this
is so that this name doesn't end up in a customer-facing location that
may cause us to violate Redis's copyright

also adds acceptance test dependency on installDist task. Previously there was not an explicit dependency so tests would fail if they happened to start before the installDist task ran