Recover paralell gateway-sender queue in stopped state
This commit implements scenario when paralell gateway-sender
is recovered in stopped state after the restart of server. Basically,
the gateway-sender is brought up to the same state as it would be
after it is started and then stopped using gfsh commands. This
way gateway-sender queue region doesn't block the collocated data
region to reach online status.

Additionally, new dunit test are introduced to verify above behavior:
- ParallelGatewaySenderOperationClusterConfigDUnitTest
- SerialGatewaySenderOperationClusterConfigDUnitTest