GEODE-8320: Set a higher value for GET_TRANSACTION_EVENTS_FROM_QUEUE_RETRIES to fix flakyness in test cases
After GEODE-8971 (Add grace period to stop gateway sender when group-transaction-events
is enabled), the value for GET_TRANSACTION_EVENTS_FROM_QUEUE_RETRIES was reduced
from 11 to 2 given that a sleep of 1 second was added between retries.
This may be the cause of sometimes seeing the
SerialWANStatsDUnitTest.testReplicatedSerialPropagationHAWithGroupTransactionEvents
fail.

The number of retries has been increased to 5 with the hope that
the test case will not fail again.