GEODE-9141: Hang while shutting down a cache server due to corrupted message
The corruption is occuring because another thread was using it to
perform a handshake on a new Connection and was still in
SocketChannel.read().  The thread initiating the handshake was
interrupted by CacheServer.close() and released the buffer prematurely,
allowing another thread to grab the still in-use buffer and serialize
a message into it.  The handshake thread's SocketChannel.read()
overwrites some of the content of the buffer before it is transmitted.

The fix is to include the input buffer as state that needs to be
protected by the stateLock and allow the handshake thread to retain
ownership of the buffer, releasing it in it's cleanup code.

I also deleted some very old checks for an IBM JRE that is no longer
required and was interfering with efforts to write a unit test.