Tweaking test based on feedback and fixing failures
Added a new interface for the AddEntryFunction to describe what
the function should do.

Cleaned up some string generation code so functions could be smaller.

Increasing the tolerance in the measurements up to 7 bytes. It seems we don't
always get the exact same value.

Trying to fix the JDK 11 test failure by excluding JDK classes from our size
measurement. This is just the JDK classes, not objects that are instances of
those classes.