Minor improvment and fix for TC's that result with timeout exceeded
Improves solution related to recovery in stopped state. Parallel gateway-sender
queue will be created/recovered at gateway-sender creation only if colocated data
region is also available. If data region is not available, then parallel gateway
sender queue will be created/recovered after colocated data region is created/recovered.

Fixed TC: testpersistentWanGateway_restartSender_expectAllEventsReceived_scenario2

This TC was failing because gateway-sender queue persistent partition region
is destroyed on all members when gateway-sender is started with clean queues,
and due to that events were not replicated to the other site. This happened because
now gateway-sender partition persistent queue is recovered/created when manual-start
set to true during server startup. The reason why this worked before this improvement
is  because partition region was not automatically created/recovered during server startup,
but only after gateway-sender is started with start command. So in that case when
"start gateway-sender with clean queue" is executed, the queue region is not destroyed,
since it doesn't exist, and therefore queued data remains on the other members.
I have changed this case to reflect the new behavior, and to check that everything works
as expected after gateway-sender is started with clean queues.

other TC's:

The other TC's were failing because partitioned persistent regions were created
in synchronous way. Now partition persistent queue region is created/recovered
during the creation of data region, so it is now necessary that partition region
is created in asynchronous way, so that members don't infinitely wait for member
with most up to date persistent data.