GEODE-9217: Adding memory overhead tests for redis (#6405)
Adding integration tests that measure the memory overhead of both native redis
and radish.  These tests measure the memory used before and after adding
entries. They assert that the measured per entry overhead is within 7 bytes of
the expected per entry overhead. They are measuring the all of the overhead -
whether it's from geode region entries or our radish data structures.

To get these tests to work on JDK 11, we added a few more --add-opens
to test.gradle. The ObjectGraphSizer was also fixed to exclude PhantomReferences,
which JDK 11 now uses in DirectByteBuffers.