Changing the constants in RedisHash to match the new overhead
Now that we are using Object2ObjectOpenHashMap, the per field overhead of the
RedisHash has changed.

Altering the test of the values we use for the constants in RedisHash. We were
previously computing the per field overhead for very small hashes and asserting
that the constants exactly match those numbers.

The new test instead computes an *average* per field overhead, and asserts that
the constant matches the average overhead.

This allows us to have a constant that provides a better estimate with
different numbers of fields. To reflect this, the tolerance in the later tests
that compare measured vs. computed overhead is now 3% instead of 5%.