GEODE-9265: Introduce redis SlotAdvisor to manage slot information
We're currently using a function to pull bucket/slot info from each
member. However, members already contain the most current view of bucket
locations so we should use that instead. This will avoid an unnecessary
fn call as well as providing more up-to-date information.

This commit introduces a `SlotAdvisor` which is intended to be the
central point of information regarding buckets/slots/members. This will
also be used to support the MOVED response.