GEODE-9223: Removing ByteArrayWrapper from hashes  (#6434)
No longer keeping ByteArrayWrapper classes in memory for RedisHashes or
creating ByteArrayWrapper classes when performing hash operations.

The main change to the RedisHash data structure was to switch from a
HashMap<ByteArrayWrapper, ByteArrayWrapper> to a
Object2ObjectOpenCustomHashMap<byte[], byte[]>. This saves ~50 bytes for each
field in the hash to not have the wrappers. Removing the extra allocations when
performing hash operations also helps performance.

Also removing a couple of extra unneeded allocations during logging and message
parsing. We should always create ArrayLists, etc. with a size if it is known so
those objects don't have to resize during parsing/deserialization.

Moving the hash commands executor to a shared object rather than allocating it
for each operation, which is extra GC pressure we don't need.

Altering the test of the values we use for the constants in RedisHash. We were
previously computing the per field overhead for very small hashes and asserting
that the constants exactly match those numbers.

The new test instead computes an *average* per field overhead, and asserts that
the constant matches the average overhead.

This allows us to have a constant that provides a better estimate with
different numbers of fields. To reflect this, the tolerance in the later tests
that compare measured vs. computed overhead is now 3% instead of 5%.

These changes BREAK compatibility with geode 1.14. The serialization format for
RedisHash, AddsDeltaInfo, and RemsDeltaInfo has changed to not write
ByteArrayWrapper classes on the wire.